OPENEUROPA-2310: Small behat context refactor.
