+++ behat.yml.dist
@@ -10,6 +10,7 @@ default:
         - Drupal\DrupalExtension\Context\MessageContext
         - Drupal\Tests\oe_media\Behat\DrupalContext
         - Drupal\Tests\oe_media\Behat\AvPortalContext
+        - Drupal\Tests\oe_media\Behat\MediaContext
         - Drupal\Tests\oe_media\Behat\CleanupContext
         - Drupal\Tests\oe_media\Behat\WebtoolsContext
         - OpenEuropa\Behat\TransformationContext:
